requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
openai>=1.30.0
python-dotenv>=1.0.0

# ── Ollama (LangChain wrapper) ───────────────────────────────────────────────
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# ── Data / numerics ──────────────────────────────────────────────────────────
numpy>=1.24.0

# ── Visualisation ────────────────────────────────────────────────────────────
matplotlib>=3.7.0
